+++ x/supply/exported/exported.go
@@ -8,7 +8,7 @@ import (
 
 // ModuleAccountI defines an account interface for modules that hold tokens in an escrow
 type ModuleAccountI interface {
-	exported.Account
+	exported.AccountI
 
 	GetName() string
 	GetPermissions() []string